backend/packages/web/pyproject.toml
[project]
name = "web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.10",
    "lib",
    "pyjwt>=2.10.1",
    "python-slugify>=8.0.4",
    "alembic>=1.14.1",
]

[tool.uv.sources]
lib = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
